Migrate scheduler implementations to use centralized ifex-scheduler library
Phase 1 & 2: Vehicle Sync Bridge and Cloud Test Service
- Replace SyncedJobState::ComputeHash() with sched_lib::compute_job_content_hash()
- Replace ComputeStateChecksumXxHash() with sched_lib::compute_state_checksum()
- Remove legacy CRC32 ComputeStateChecksum() function
- Update GetStateChecksum() to return uint64_t (xxHash64)
- Update tests to use uint64_t checksum type
- Add ifex-scheduler library dependency to CMakeLists.txt
- Update comments to reflect library usage

The ifex-scheduler library provides canonical hash computation ensuring
all implementations produce identical checksums for sync protocol v2.

+++ cloud/cloud-scheduler-service/tests/cloud_scheduler_integration_test.cpp
@@ -1349,12 +1349,12 @@ TEST_F(SchedulerBidirectionalSyncTest, QuiescenceChecksumConsistency) {
     // When no changes occur, the checksum should remain stable.
 
     // Get initial checksum
-    uint32_t checksum1 = vehicle_sync_bridge_->GetStateChecksum();
+    uint64_t checksum1 = vehicle_sync_bridge_->GetStateChecksum();
     LOG(INFO) << "Initial checksum: " << checksum1;
 
     // Wait and check again - should be identical
     std::this_thread::sleep_for(1s);
-    uint32_t checksum2 = vehicle_sync_bridge_->GetStateChecksum();
+    uint64_t checksum2 = vehicle_sync_bridge_->GetStateChecksum();
     LOG(INFO) << "Checksum after wait: " << checksum2;
 
     EXPECT_EQ(checksum1, checksum2)
@@ -1382,7 +1382,7 @@ TEST_F(SchedulerBidirectionalSyncTest, QuiescenceChecksumConsistency) {
     // Wait for sync
     std::this_thread::sleep_for(3s);
 
-    uint32_t checksum3 = vehicle_sync_bridge_->GetStateChecksum();
+    uint64_t checksum3 = vehicle_sync_bridge_->GetStateChecksum();
     LOG(INFO) << "Checksum after job added: " << checksum3;
 
     // Checksum should change after adding a job
